refactor: LinkButton 사용
