+++ src/components/login/Login.tsx
@@ -1,6 +1,5 @@
-import { Link } from 'react-router-dom'
 import { Button, Paper, Stack, Text } from '@mantine/core'
-import { Logo } from 'components/common'
+import { LinkButton, Logo } from 'components/common'
 import { CredentialInput } from './CredentialInput'
 import { paths } from 'routes'
 
@@ -23,9 +22,7 @@ export const OAuthLogin = () => (
 export const TryRegister = () => (
   <>
     <Text>아직 계정이 없으신가요?</Text>
-    <Button component={Link} to={paths.register} variant='default'>
-      회원가입
-    </Button>
+    <LinkButton to={paths.register}>회원가입</LinkButton>
   </>
 )
